finished testing for the active customer model
