--- app/activeCustomer.js
@@ -1,11 +1,0 @@
-'use strict';
-
-let activeCustomer = {
-  id: null
-}
-
-module.exports.setActiveCustomer = (id) => {
-  activeCustomer.id = id;
-}
-
-module.exports.getActiveCustomer = () => activeCustomer;
