+++ app/controllers/activeCustomerCtrl.js
@@ -1,0 +1,12 @@
+// 'use strict';
+
+// const { setActiveCustomer, getActiveCustomer } = require('../models/activeCustomers');
+
+// module.exports.activeCustomer = (req, res, next) => {
+//     setActiveCustomers()
+//     .then( (data) => {
+//       res.status(200).json(data);
+//     })
+//     .catch( (err) => next(err));
+//   }
+// };
